refactor(vats): factor out lib-boot.js
